% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_file_type.R
\name{check_for_app_file}
\alias{check_for_app_file}
\title{Get the file type a shiny app directory}
\usage{
check_for_app_file(app_loc, error_on_missing = FALSE)
}
\arguments{
\item{app_loc}{Path to a shiny app}

\item{error_on_missing}{Should the lack of
app ui file trigger an error? If not returns a type of "missing" and no path}
}
\value{
either \code{TRUE} if it finds an (\code{app.R}) or \code{FALSE} if no app detected
}
\description{
Also checks for multifile apps and emits a depreciation error
}
\keyword{internal}
