% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_ui_fn.R
\name{deparse_ui_fn}
\alias{deparse_ui_fn}
\title{Deparse ui function tree to expression}
\usage{
deparse_ui_fn(ui_tree, remove_namespace = FALSE)
}
\arguments{
\item{ui_tree}{Valid ui tree intermediate representation of an apps ui}

\item{remove_namespace}{Should the generated code have the namespaces removed
or should generated function calls take the form of \code{pkg::fn()}}
}
\value{
A list with \code{call}: language expression of the generated code, and
\code{namespaces_removed}: a character vector of all the namespaces that were
stripped from the ui functions (only has elements if \code{remove_namespaces = TRUE})
}
\description{
Deparse ui function tree to expression
}
