% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_app_template.R
\name{generate_app_template_files}
\alias{generate_app_template_files}
\title{Generate code for app files from a template}
\usage{
generate_app_template_files(app_template, remove_namespace = TRUE)
}
\arguments{
\item{app_template}{Template object. See details for format.}

\item{remove_namespace}{}
}
\value{
A list with either a single "app_file" field on it when a single-file
app has been requested, or both a "ui_file" and "server_file" on it for
both scripts of a multi-file app
}
\description{
Generate code for app files from a template
}
\details{
Template object
The app template has the following information attached to it
uiTree: ui AST node;
outputType: "single-file" | "multi-file"
otherCode: {
# Extra code that will be coppied unchanged above the ui definition
uiExtra?: string;

\if{html}{\out{<div class="sourceCode">}}\preformatted{# List of libraries that need to be loaded in server code
serverLibraries?: string[];

# Extra code that will be copied unchanged above server
# function definition
serverExtra?: string;

# Body of server function. This will be wrapped in the code
# `function(input, output)\{....\}`
serverFunctionBody?: string;
}\if{html}{\out{</div>}}

};
}
\keyword{internal}
