% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_file_type.R
\name{get_app_file_type}
\alias{get_app_file_type}
\title{Get the file type a shiny app directory}
\usage{
get_app_file_type(app_loc, error_on_missing = FALSE)
}
\arguments{
\item{app_loc}{Path to a shiny app}

\item{error_on_missing}{Should the lack of
app ui file trigger an error? If not returns a type of "missing" and no path}
}
\value{
either "SINGLE-FILE" (\verb{app.R``), "MULTI-FILE" (}ui.R\code{and}server.R`), or "MISSING" (empty directory)
}
\description{
Get the file type a shiny app directory
}
\keyword{internal}
