% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_info.R
\name{get_app_info}
\alias{get_app_info}
\title{Get info about Shiny app for editor}
\usage{
get_app_info(app_loc)
}
\arguments{
\item{app_loc}{Path the the app location}
}
\value{
If it's a
single file app its a list with \code{app_type = "SINGLE-FILE"} and a property
\code{app} that contains \code{script} (the raw file text for \code{app.R}) and \code{ast} which
is the serialized ast for that script.

If it's a multi-file app its \verb{type="MULTI-FILE} and \code{ui}  and \code{server}
properties with the same script + ast combo as \code{app} in the single-file case
}
\description{
Gets pointed at an app containing directory and returns one of two
datastructures depending on if it's a single- or multi-file app.
}
\keyword{internal}
