% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_file_helpers.R
\name{get_app_ui_file}
\alias{get_app_ui_file}
\title{Get the file defining the ui for a shiny app directory}
\usage{
get_app_ui_file(app_loc, error_on_missing = FALSE)
}
\arguments{
\item{app_loc}{Path to a shiny app @param error_on_missing Should the lack of
app ui file trigger an error? If not returns a type of "missing" and no path}
}
\value{
Path to either the \code{app.R} file in directory in the case of
single-file apps, or \code{ui.R} in the case of multi-file apps. If both exist
then \code{app.R} will take precedence
}
\description{
Get the file defining the ui for a shiny app directory
}
\keyword{internal}
