% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ui_definition_info.R
\name{get_loaded_libraries}
\alias{get_loaded_libraries}
\title{Grab libraries loaded via library() calls}
\usage{
get_loaded_libraries(file_lines)
}
\arguments{
\item{file_lines}{lines of an R scripto}
}
\value{
character vector of libraries loaded with library() call
}
\description{
Gather the name of all libraries loaded with calls to \code{library()} in a given
app
}
\examples{
file_lines <- c(
  "library(gridlayout) ",
  "# hi there I'm a comment",
  "   library(testing)",
  "#library(commentedOut)",
  "ui <- grid_page()"
)

shinyuieditor:::get_loaded_libraries(file_lines)

}
\keyword{internal}
