% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudioapi_utils.R
\name{get_path_to_app_server_script}
\alias{get_path_to_app_server_script}
\title{Get path of the app server script (either \code{app.R} or \code{server.R})}
\usage{
get_path_to_app_server_script(app_type, app_loc)
}
\arguments{
\item{app_type}{Type of app to get the server script for. Can be
"SINGLE-FILE" or "MULTI-FILE". As returned by \code{get_app_file_type()}}

\item{app_loc}{Path to directory containing Shiny app to be visually edited
(either containing an \code{app.R} or both a \code{ui.R} and \code{server.R}).}
}
\description{
Get path of the app server script (either \code{app.R} or \code{server.R})
}
\keyword{internal}
