% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudioapi_utils.R
\name{insert_server_code}
\alias{insert_server_code}
\title{Insert code into the server script at the given location}
\usage{
insert_server_code(snippet, insert_at, app_loc, app_type)
}
\arguments{
\item{snippet}{Code to be inserted}

\item{insert_at}{Location to insert the code at}

\item{app_loc}{Path to directory containing Shiny app to be visually edited
(either containing an \code{app.R} or both a \code{ui.R} and \code{server.R}).}

\item{app_type}{Type of app to get the server script for. Can be
"SINGLE-FILE" or "MULTI-FILE". As returned by \code{get_app_file_type()}}
}
\description{
Insert code into the server script at the given location
}
\keyword{internal}
