% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_known_ui_fn.R
\name{is_known_ui_fn}
\alias{is_known_ui_fn}
\title{Is a function call a known UI function call?}
\usage{
is_known_ui_fn(x)
}
\arguments{
\item{x}{Language object node corresponding to a function call}
}
\value{
Boolean
}
\description{
Takes an AST node and tells us if it corresponds to a known UI function call.
Makes the assumption that we can't look into the environment to make sure
that we are getting the function we think we are. This is because we want to
be able to parse raw text instead of relying on an active evaulation context.
}
\examples{

is_known_ui_fn(rlang::expr(gridlayout::grid_page()))
is_known_ui_fn(rlang::expr(custom_ui_page()))

}
