% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_editor.R
\name{launch_editor}
\alias{launch_editor}
\title{Launch visual editor server}
\usage{
launch_editor(ui_loc, port = 8888, show_logs = TRUE)
}
\arguments{
\item{ui_loc}{Path to the \code{ui.R} file containing the ui to be edited. If file
does not exist a default starter template will be used upon finishing will
be saved to the path.}

\item{port}{Port to launch server on. Should only be changed for dev work.}

\item{show_logs}{Print status messages to the console? For debugging.}
}
\value{
An \code{httpuv} app server object (as returned by \code{httpuv::startServer}).
To terminate before finishing with the app run \code{s$stop()} (assuming \code{s} is
the return value of this function.)
}
\description{
Spins up an \code{httpuv} server that handles parsing and deparsing of the UI tree
and saving.
}
