% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tree_to_code.R
\name{ui_tree_to_code}
\alias{ui_tree_to_code}
\title{Convert ui tree IR to code text}
\usage{
ui_tree_to_code(ui_tree, remove_namespace = TRUE)
}
\arguments{
\item{ui_tree}{Valid ui tree intermediate representation of an apps ui}

\item{remove_namespace}{Should the generated code have the namespaces removed
or should generated function calls take the form of \code{pkg::fn()}}
}
\value{
A list with \code{text}: lines of the generated code, and
\code{namespaces_removed}: a character vector of all the namespaces that were
stripped from the ui functions (only has elements if \code{remove_namespaces = TRUE})
}
\description{
Convert ui tree IR to code text
}
\keyword{internal}
