% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ui_definition_info.R
\name{update_ui_definition}
\alias{update_ui_definition}
\title{Update the ui definition for a given ui-defining file}
\usage{
update_ui_definition(file_info, new_ui_tree, remove_namespace = TRUE)
}
\arguments{
\item{file_info}{Info on the ui-defining file as obtained by
\code{shinyuieditor:::get_file_ui_definition_info}}

\item{new_ui_tree}{The new UI IR tree defining the new ui for the file}

\item{remove_namespace}{Should the new ui be generated with namespaces
stripped and \code{library()} calls added for any non-defined namespaces?}
}
\value{
A new character vector containing the lines of the ui-defining file
with the layout updated to match the \code{new_ui_tree}.
}
\description{
Update the ui definition for a given ui-defining file
}
