% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_for_app_close.R
\name{watch_for_app_close}
\alias{watch_for_app_close}
\title{Create a watcher that checks when an app is as indicated by a websocket
connection being severed}
\usage{
watch_for_app_close(on_close)
}
\arguments{
\item{on_close}{A function to call when the app
closes @return A list of functions to call when the app opens or closes}
}
\description{
Create a watcher that checks when an app is as indicated by a websocket
connection being severed
}
\keyword{internal}
