% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_app_file}
\alias{write_app_file}
\title{Write app script to a file}
\usage{
write_app_file(app_lines, app_loc, file_type)
}
\arguments{
\item{app_lines}{Character vector containing the code for the given script.
Will be concatinated with new lines}

\item{app_loc}{Location of folder where script will be written to}

\item{file_type}{Type of file being written. Can either be "app" for writing
an "app.R", or "ui"/"server" for writing the two scripts of a multi-file
app.}
}
\description{
Write app script to a file
}
\keyword{internal}
