% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_app_template.R
\name{write_app_template}
\alias{write_app_template}
\title{Setup app from template}
\usage{
write_app_template(app_template, app_loc, remove_namespace = TRUE)
}
\arguments{
\item{app_template}{Template object. See details for format.}

\item{app_loc}{Path to directory containing Shiny app to be visually edited
(either containing an \code{app.R} or both a \code{ui.R} and \code{server.R}). If the
provided location doesn't exist, or doesn't contain an app, the user will
be prompted to choose a starter template which will then be written to the
location specified.}

\item{remove_namespace}{Should the generated code have the namespaces removed
or should generated function calls take the form of \code{pkg::fn()}}
}
\description{
Takes a template object as provided by the front-end and generates a single
or multi-file app in provided location.
}
\keyword{internal}
