% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_ui_file.R
\name{get_app_ui_file}
\alias{get_app_ui_file}
\title{Get the file defining the ui for a shiny app directory}
\usage{
get_app_ui_file(app_loc)
}
\arguments{
\item{app_loc}{Path to a shiny app}
}
\value{
Path to either the \code{app.R} file in directory in the case of
single-file apps, or \code{ui.R} in the case of multi-file apps. If both exist
then \code{app.R} will take precedence
}
\description{
Get the file defining the ui for a shiny app directory
}
