% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ui_definition_info.R
\name{get_file_ui_definition_info}
\alias{get_file_ui_definition_info}
\title{Get info about ui definition from file}
\usage{
get_file_ui_definition_info(file_lines, type = "single-file")
}
\arguments{
\item{file_lines}{Character vector of the lines of the file that defines a
shiny app's ui (as from \code{readLines()}).}

\item{type}{Is the app a single-file app? E.g. app is container entirely in
\code{app.R}? Or is it \code{multi-file}?}
}
\value{
List with both the \code{type} and \code{file_lines} mirrored from the
arguments of the same name. Along with this the \code{ui_bounds} containing the
start and end lines of the file section that defines the app's ui, the
\code{ui_tree} IR that defines that ui, and \code{loaded_libraries}: libraries loaded
via \code{library()} calls in the script.
}
\description{
Get info about ui definition from file
}
