% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_editor.R
\name{launch_editor}
\alias{launch_editor}
\title{Launch visual editor server}
\usage{
launch_editor(
  ui_loc,
  host = "127.0.0.1",
  port = 8888,
  shiny_background_port = 4444,
  show_logs = TRUE,
  run_in_background = FALSE
)
}
\arguments{
\item{ui_loc}{Path to the \code{ui.R} file containing the ui to be edited. If file
does not exist a default starter template will be used upon finishing will
be saved to the path.}

\item{host}{A string that is a valid IPv4 address that is owned by this
server, or \code{"0.0.0.0"} to listen on all IP addresses.}

\item{port}{A number or integer that indicates the server port that should be
listened on. Note that on most Unix-like systems including Linux and Mac OS
X, port numbers smaller than 1025 require root privileges.}

\item{shiny_background_port}{Port to launch the running app preview on. Again only used for dev work.}

\item{show_logs}{Print status messages to the console? For debugging.}

\item{run_in_background}{Should the app run in a background process or block
the console? See \code{?httpuv::startServer()} vs \code{?httpuv::runServer()}.}
}
\value{
An \code{httpuv} app server object (as returned by \code{httpuv::startServer}).
To terminate before finishing with the app run \code{s$stop()} (assuming \code{s} is
the return value of this function.)
}
\description{
Spins up an \code{httpuv} server that handles parsing and deparsing of the UI tree
and saving.
}
