% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ui_fn.R
\name{parse_ui_fn}
\alias{parse_ui_fn}
\title{Parse UI function node}
\usage{
parse_ui_fn(ui_node_expr, env = rlang::caller_env())
}
\arguments{
\item{ui_node_expr}{A language object representing the call of a known Shiny
UI function.}

\item{env}{Environment in which to evaluate arguments. Necessary for filling
in the names of all arguments. Almost always left as the default.}
}
\value{
A nested list describing the UI of the app for use in the Shiny
Visual Editor
}
\description{
Function to recursively parse a Shiny UI definition. Will build a nested list
of known UI functions and their arguments.
}
\examples{

app_expr <- rlang::expr(
  gridlayout::grid_page(
    layout = "
            | 1rem | 250px   | 1fr  |
            |------|---------|------|
            | 1fr  | sidebar | plot |",
    gridlayout::vertical_stack_panel(
      area = "sidebar",
      item_alignment = "center",
      shiny::sliderInput(
        inputId = "bins",
        label = "Num Bins",
        min = 10L,
        max = 100L,
        value = 40L
      )
    ),
    gridlayout::vertical_stack_panel(
      area = "plot",
      item_alignment = "center",
      shiny::plotOutput(
        outputId = "distPlot",
        height = "100\%"
      )
    )
  )
)
lobstr::tree(parse_ui_fn(app_expr))

}
