% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ui_fn.R
\name{parse_ui_fn}
\alias{parse_ui_fn}
\title{Parse UI function node}
\usage{
parse_ui_fn(ui_node_expr)
}
\arguments{
\item{ui_expr}{A language object representing the call of a known Shiny UI function.}
}
\value{
A nested list describing the UI of the app for use in the Shiny Visual Editor
}
\description{
Function to recursively parse a Shiny UI definition. Will build a nested list of known UI functions and their arguments.
}
\examples{

app_expr <- rlang::expr(
  gridlayout::grid_page(
    layout = gridlayout::new_gridlayout("
      |2rem  |200px   |1fr    |
      |80px  |header  |header |
      |1fr   |sidebar |plot   |"),
    theme = bslib::bs_theme(),
    header = gridlayout::title_panel("This is my header"),
    sidebar = gridlayout::grid_panel(
      title = "Settings",
      shiny::sliderInput("bins", "Number of bins:", min = 1, value = 5, max = 10)
    ),
    plot = shiny::plotOutput("distPlot")
  )
)


lobstr::tree(parse_ui_fn(app_expr))
}
