% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_app_template.R
\name{write_app_template}
\alias{write_app_template}
\title{Setup app from template}
\usage{
write_app_template(app_template, loc, remove_namespace = TRUE)
}
\arguments{
\item{app_template}{Template object. See details for format.}

\item{loc}{Location to place app template}

\item{remove_namespace}{Should the generated code have the namespaces removed
or should generated function calls take the form of \code{pkg::fn()}}
}
\description{
Takes a template object as provided by the front-end and generates a single
or multi-file app in provided location.
}
\details{
Template object
The app template has the following information attached to it
uiTree: ui AST node;
outputType: "single-file" | "multi-file"
otherCode: {
# Extra code that will be coppied unchanged above the ui definition
uiExtra?: string;

\if{html}{\out{<div class="sourceCode">}}\preformatted{# List of libraries that need to be loaded in server code
serverLibraries?: string[];

# Extra code that will be copied unchanged above server
# function definition
serverExtra?: string;

# Body of server function. This will be wrapped in the code
# `function(input, output)\{....\}`
serverFunctionBody?: string;
}\if{html}{\out{</div>}}

};
}
\keyword{internal}
